Update io-network.rst to
I know this is picky, but the documentation for open(file_name, mode) doesn't explain that file_name and mode should be strings. The file_name is fairly obvious, but one might be wonder whether mode should be "r" or 'r' or even r (e.g., a reserved word). The fact that it is string-based is noted in the explanation, but the table does not show that. The proposed change makes explicit that file_name should be a string, and lists possible modes as strings in the table.